requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
banks==2.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
executing==2.2.1
filelock==3.24.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ipykernel==7.2.0
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.14
llama-index-cli==0.5.3
llama-index-core==0.14.14
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-huggingface-api==0.6.2
llama-index-llms-openai==0.6.18
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.14.2
llama-parse==0.6.54
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx
nltk
numpy
openai
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt_toolkit
propcache
psutil
ptyprocess
pure_eval
pydantic
pydantic_core
Pygments
pypdf
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
regex
requests
rich
shellingham
six
sniffio
soupsieve
SQLAlchemy
stack-data
striprtf
tenacity
tiktoken
tornado
tqdm
traitlets
typer
typer-slim
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
wcwidth
wrapt
yarl